Move the file from main.go to the appropriate file
